+++ src/onegov.winterthur/onegov/winterthur/app.py
@@ -11,7 +11,7 @@ from onegov.winterthur.theme import WinterthurTheme
 class WinterthurApp(OrgApp):
 
     #: the version of this application (do not change manually!)
-    version = '0.2.6'
+    version = '0.2.7'
 
     frame_ancestors = {
         'https://winterthur.ch',