+++ src/onegov.winterthur/onegov/winterthur/theme/winterthur_theme.py
@@ -14,7 +14,7 @@ class WinterthurTheme(OrgTheme):
 
     # don't touch this number, it's incremented using bumpversion, so every
     # release will automatically trigger a rebuild of the theme
-    version = '0.2.6'
+    version = '0.2.7'
 
     @property
     def post_imports(self):